+++ flint/prefect/flows/continuum_pipeline.py
@@ -54,6 +54,7 @@ from flint.prefect.common.imaging import (
     task_wsclean_imager,
     task_zip_ms,
 )
+from flint.prefect.common.ms import task_add_model_source_list_to_ms
 from flint.prefect.common.utils import (
     task_archive_sbid,
     task_create_beam_summary,
@@ -303,11 +304,20 @@ def process_science_fields(
         round_info="initial",
     )  # type: ignore
 
+    wsclean_cmds = (
+        task_add_model_source_list_to_ms.map(
+            wsclean_command=wsclean_cmds,
+            calibrate_container=field_options.calibrate_container,
+        )
+        if field_options.update_model_data_with_source_list
+        else wsclean_cmds
+    )
+
     # TODO: This should be waited!
     beam_summaries = task_create_beam_summary.map(
         ms=preprocess_science_mss, imageset=wsclean_cmds
     )
-
+    archive_wait_for.extend(beam_summaries)
     archive_wait_for.extend(wsclean_cmds)
 
     beam_aegean_outputs = None
@@ -420,6 +430,14 @@ def process_science_fields(
                 mode="wsclean",
                 round_info=current_round,
             )  # type: ignore
+            wsclean_cmds = (
+                task_add_model_source_list_to_ms.map(
+                    wsclean_command=wsclean_cmds,
+                    calibrate_container=field_options.calibrate_container,
+                )
+                if field_options.update_model_data_with_source_list
+                else wsclean_cmds
+            )
             archive_wait_for.extend(wsclean_cmds)
 
             # Do source finding on the last round of self-cal'ed images
@@ -497,7 +515,7 @@ def process_science_fields(
 
     # zip up the final measurement set, which is not included in the above loop
     if field_options.zip_ms:
-        task_zip_ms.map(in_item=wsclean_cmds)
+        task_zip_ms.map(in_item=wsclean_cmds, wait_for=archive_wait_for)
 
     if field_options.sbid_archive_path or field_options.sbid_copy_path and run_aegean:
         update_archive_options = get_options_from_strategy(