+++ tests/test_ms.py
@@ -6,18 +6,20 @@ import shutil
 from pathlib import Path
 
 import numpy as np
+from pydantic import ValidationError
 import pytest
 from casacore.tables import table
 
 from flint.calibrate.aocalibrate import ApplySolutions
-from flint.exceptions import MSError
 from flint.ms import (
     MS,
     check_column_in_ms,
     copy_and_preprocess_casda_askap_ms,
     find_mss,
     get_phase_dir_from_ms,
+    remove_columns_from_ms,
     rename_ms_and_columns_for_selfcal,
+    subtract_model_from_data_column,
 )
 from flint.utils import get_packaged_resource_path
 
@@ -213,10 +215,155 @@ def test_ms_from_options():
 
 
 def test_raise_error_ms_from_options():
+    """Ensure validation error is raised when not casting correct
+    ms path type"""
+    # TODO: remove this test as pydantic prevents this from happening
     path = Path("example.ms")
-    solutions = ApplySolutions(
-        cmd="none", solution_path=Path("example_sols.bin"), ms=path
+    with pytest.raises(ValidationError):
+        _ = ApplySolutions(cmd="none", solution_path=Path("example_sols.bin"), ms=path)
+
+    # with pytest.raises((MSError, ValidationError)):
+    #     MS.cast(solutions)
+
+
+@pytest.fixture
+def ms_remove_example(tmpdir):
+    """Create a copy of the MS that will be modified."""
+    ms_zip = Path(
+        get_packaged_resource_path(
+            package="flint.data.tests",
+            filename="SB39400.RACS_0635-31.beam0.small.ms.zip",
+        )
+    )
+    outpath = Path(tmpdir) / "Removecols_39400"
+
+    shutil.unpack_archive(ms_zip, outpath)
+
+    ms_path = Path(outpath) / "SB39400.RACS_0635-31.beam0.small.ms"
+
+    return ms_path
+
+
+def _get_column_names(ms_path):
+    with table(str(ms_path)) as tab:
+        column_names = tab.colnames()
+
+    return column_names
+
+
+def test_remove_columns_from_ms(ms_remove_example):
+    """Load an example MS to remove columns from"""
+    original_columns = _get_column_names(ms_path=ms_remove_example)
+
+    removed_columns = remove_columns_from_ms(
+        ms=ms_remove_example, columns_to_remove="DATA"
     )
 
-    with pytest.raises(MSError):
-        MS.cast(solutions)
+    updated_columns = _get_column_names(ms_path=ms_remove_example)
+    diff = set(original_columns) - set(updated_columns)
+    assert len(diff) == 1
+    assert list(diff)[0] == "DATA"
+    assert removed_columns[0] == "DATA"
+    assert len(removed_columns) == 1
+
+    removed_columns = remove_columns_from_ms(
+        ms=ms_remove_example, columns_to_remove="DATA"
+    )
+    assert len(removed_columns) == 0
+
+
+@pytest.fixture
+def casda_taql_example(tmpdir):
+    ms_zip = Path(
+        get_packaged_resource_path(
+            package="flint.data.tests",
+            filename="scienceData.EMU_0529-60.SB50538.EMU_0529-60.beam08_averaged_cal.leakage.ms.zip",
+        )
+    )
+    outpath = Path(tmpdir) / "taqlsubtract"
+
+    shutil.unpack_archive(ms_zip, outpath)
+
+    ms_path = (
+        Path(outpath)
+        / "scienceData.EMU_0529-60.SB50538.EMU_0529-60.beam08_averaged_cal.leakage.ms"
+    )
+
+    return ms_path
+
+
+def test_subtract_model_from_data_column(casda_taql_example):
+    """Ensure we can subtact the model from the data via taql"""
+    ms = Path(casda_taql_example)
+    assert ms.exists()
+    ms = MS(path=ms)
+
+    from casacore.tables import maketabdesc, makearrcoldesc
+
+    with table(str(ms.path), readonly=False) as tab:
+        data = tab.getcol("DATA")
+        ones = np.ones_like(data, dtype=data.dtype)
+
+        tab.putcol(columnname="DATA", value=ones)
+
+        if "MODEL_DATA" not in tab.colnames():
+            coldesc = tab.getdminfo("DATA")
+            coldesc["NAME"] = "MODEL_DATA"
+            tab.addcols(
+                maketabdesc(makearrcoldesc("MODEL_DATA", 0.0 + 0j, ndim=2)), coldesc
+            )
+            tab.flush()
+        tab.putcol(columnname="MODEL_DATA", value=ones)
+        tab.flush()
+
+    ms = subtract_model_from_data_column(
+        ms=ms, model_column="MODEL_DATA", data_column="DATA"
+    )
+    with table(str(ms.path)) as tab:
+        data = tab.getcol("DATA")
+        assert np.all(data == 0 + 0j)
+
+
+def test_subtract_model_from_data_column_ms_column(tmpdir):
+    """Ensure we can subtact the model from the data via taql"""
+    ms_zip = Path(
+        get_packaged_resource_path(
+            package="flint.data.tests",
+            filename="scienceData.EMU_0529-60.SB50538.EMU_0529-60.beam08_averaged_cal.leakage.ms.zip",
+        )
+    )
+    outpath = Path(tmpdir) / "taqlsubtract2"
+
+    shutil.unpack_archive(ms_zip, outpath)
+
+    ms_path = (
+        Path(outpath)
+        / "scienceData.EMU_0529-60.SB50538.EMU_0529-60.beam08_averaged_cal.leakage.ms"
+    )
+
+    ms = Path(ms_path)
+    assert ms.exists()
+    ms = MS(path=ms, column="DATA")
+
+    from casacore.tables import maketabdesc, makearrcoldesc
+
+    with table(str(ms.path), readonly=False) as tab:
+        data = tab.getcol("DATA")
+        ones = np.ones_like(data, dtype=data.dtype)
+
+        tab.putcol(columnname="DATA", value=ones)
+
+        if "MODEL_DATA" not in tab.colnames():
+            coldesc = tab.getdminfo("DATA")
+            coldesc["NAME"] = "MODEL_DATA"
+            tab.addcols(
+                maketabdesc(makearrcoldesc("MODEL_DATA", 0.0 + 0j, ndim=2)), coldesc
+            )
+            tab.flush()
+        tab.putcol(columnname="MODEL_DATA", value=ones)
+        tab.flush()
+
+    ms = subtract_model_from_data_column(ms=ms, model_column="MODEL_DATA")
+    with table(str(ms.path)) as tab:
+        data = tab.getcol("DATA")
+        assert np.all(data == 0 + 0j)